Internationalize browser unsupported dialog
