+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/error/error.component.spec.ts
@@ -45,13 +45,6 @@ describe('ErrorComponent', () => {
     expect(env.showDetails.style.display).toBe('none');
     expect(env.stackTrace.style.display).toBe('none');
   }));
-
-  it('should correctly generate links', fakeAsync(() => {
-    const errorComponent = new ErrorComponent(null as any, null as any);
-    expect(errorComponent.getLinkHTML('example', 'https://example.com')).toEqual(
-      `<a href="https://example.com" target="_blank">example</a>`
-    );
-  }));
 });
 
 @Directive({
