+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/error/error.component.ts
@@ -1,7 +1,6 @@
 import { MDC_DIALOG_DATA, MdcDialogRef } from '@angular-mdc/web/dialog';
 import { Component, Inject } from '@angular/core';
-import { translate } from '@ngneat/transloco';
-import { issuesEmailTemplate, supportedBrowser } from 'xforge-common/utils';
+import { browserLinks, getLinkHTML, issuesEmailTemplate, supportedBrowser } from 'xforge-common/utils';
 import { environment } from '../../environments/environment';
 
 export interface ErrorAlert {
@@ -21,22 +20,10 @@ export class ErrorComponent {
   constructor(public dialogRef: MdcDialogRef<ErrorComponent>, @Inject(MDC_DIALOG_DATA) public data: ErrorAlert) {}
 
   get issueEmailLink() {
-    return this.getLinkHTML(environment.issueEmail, issuesEmailTemplate(this.data.eventId));
+    return getLinkHTML(environment.issueEmail, issuesEmailTemplate(this.data.eventId));
   }
 
-  get chromeLink() {
-    return this.getLinkHTML(translate('error.chrome'), 'https://www.google.com/chrome/');
-  }
-
-  get firefoxLink() {
-    return this.getLinkHTML(translate('error.firefox'), 'https://firefox.com');
-  }
-
-  getLinkHTML(text: string, href: string) {
-    const a = document.createElement('a');
-    a.href = href;
-    a.setAttribute('target', '_blank');
-    a.textContent = text;
-    return a.outerHTML;
+  get browserLinks() {
+    return browserLinks();
   }
 }
