+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/supported-browsers-dialog/supported-browsers-dialog.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { browserLinks } from 'xforge-common/utils';
 
 @Component({
   selector: 'app-supported-browsers-dialog',
@@ -6,4 +7,8 @@ import { Component } from '@angular/core';
 })
 export class SupportedBrowsersDialogComponent {
   constructor() {}
+
+  get browserLinks() {
+    return browserLinks();
+  }
 }
